fix(schema): INT → BIGINT en params/returns de entidades + TIMESTAMPTZ → TIMESTAMP + shipping en run_all.sql
- 08_fin.sql: usp_bank_statementline_insert TIMESTAMPTZ → TIMESTAMP
- usp_crm.sql: p_customer_id BIGINT en lead_close, activity_list, activity_create
- usp_fleet.sql: p_supplier_id BIGINT en maintenanceorder_create
- usp_inv.sql: p_product_id/p_customer_id BIGINT + RETURNS TABLE cols BIGINT
  (BinStock_List, Lot_List, Lot_Get, Lot_Create, Serial_List, Serial_Get,
   Serial_Register, Serial_UpdateStatus, Movement_List, Movement_Create,
   Valuation_GetMethod, Valuation_SetMethod)
- usp_inv_integracion.sql: p_product_id/p_customer_id BIGINT
- usp_logistics.sql: p_supplier_id/p_customer_id BIGINT (5 funciones)
- usp_mfg.sql: p_product_id BIGINT (3 funciones)
- run_all.sql: agregar usp_shipping.sql + usp_inv_integracion.sql
- migrations/00025: DROP funciones INT antes de recrear con BIGINT en prod

+++ web/api/sqlweb-pg/includes/sp/usp_crm.sql
@@ -454,7 +454,7 @@ CREATE OR REPLACE FUNCTION usp_crm_lead_close(
     p_lead_id     INT,
     p_is_won      BOOLEAN,
     p_lost_reason VARCHAR(500) DEFAULT NULL,
-    p_customer_id INT DEFAULT NULL,
+    p_customer_id BIGINT DEFAULT NULL,
     p_user_id     INT DEFAULT NULL
 )
 RETURNS TABLE("ok" INT, "mensaje" VARCHAR)
@@ -508,7 +508,7 @@ DROP FUNCTION IF EXISTS usp_crm_activity_list(INT, INT, INT, BOOLEAN, TIMESTAMP,
 CREATE OR REPLACE FUNCTION usp_crm_activity_list(
     p_company_id   INT,
     p_lead_id      INT DEFAULT NULL,
-    p_customer_id  INT DEFAULT NULL,
+    p_customer_id  BIGINT DEFAULT NULL,
     p_is_completed BOOLEAN DEFAULT NULL,
     p_due_before   TIMESTAMP DEFAULT NULL,
     p_page         INT DEFAULT 1,
@@ -570,7 +570,7 @@ DROP FUNCTION IF EXISTS usp_crm_activity_create(INT, INT, INT, VARCHAR, VARCHAR,
 CREATE OR REPLACE FUNCTION usp_crm_activity_create(
     p_company_id         INT,
     p_lead_id            INT DEFAULT NULL,
-    p_customer_id        INT DEFAULT NULL,
+    p_customer_id        BIGINT DEFAULT NULL,
     p_activity_type      VARCHAR(30) DEFAULT NULL,
     p_subject            VARCHAR(200) DEFAULT NULL,
     p_description        TEXT DEFAULT NULL,
